Completed Player Health Bar Reference
Updated the OnGUI function to use a reference to the healthBarLength
value rather than needing to calculate it every frame.
The reference will decrease research use as it will allow for the OnGUI
function to only calculate when the value of healthBarLength actually
changes rather than constantly updating no change to the bar.