Caching Enemy Transform
Added in initial variables that will be used to control which target the
enemy will face/rotate and the speed at which it will move to the
target.
In Awake, myTransform is used to cache the value of the enemy's
Transform to maintain a reference to the data rather than needing to
constantly look up the data. Cheaper on resources.