Created Distance Check for Melee
Created a variable and Debug statement to print the distance between the
Player and the targeted object.
Used this to find the optimal distance to test whether the Player is
within Melee range of the Enemy object and if so to grab reference to
the EnemyHealth component of the Enemy object, call the
AdjustCurrentHealth function with the EnemyHealth Class and thus reduce
health by 10.