Created Facing Check for Melee
Created variables to compare the Dot product of the normalized position
of the target relative to ourselves and our forward facing direction.
Debugged to help find best value to allow for melee to take place to the
left or right of the target.
Updated check to insure that the Melee attack is taking place within a
sensible range of the target.